Move path_planner into separate file
